+++ src/Type/User.php
@@ -29,6 +29,7 @@ final readonly class User
         public ?bool $hasTopicsEnabled = null,
         public ?bool $allowsUsersToCreateTopics = null,
         public ?bool $canManageBots = null,
+        public ?bool $supportsJoinRequestQueries = null,
     ) {}
 
     public function toRequestArray(): array
@@ -52,6 +53,7 @@ final readonly class User
                 'has_topics_enabled' => $this->hasTopicsEnabled,
                 'allows_users_to_create_topics' => $this->allowsUsersToCreateTopics,
                 'can_manage_bots' => $this->canManageBots,
+                'supports_join_request_queries' => $this->supportsJoinRequestQueries,
             ],
             static fn(mixed $value): bool => $value !== null,
         );