+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - New #200: Add `richMessage` parameter to `EditMessageText` method.
 - Enh #200: Make `text` parameter of `EditMessageText` method optional.
 - New #200: Add `richMessage` field to `Message` type.
+- New #200: Add `supportsJoinRequestQueries` field to `User` type.
 
 ## 0.21 May 9, 2026